Reparametrised gamma distribution

Density, distribution function, quantile function and random generation for the gamma distribution reparametrised in terms of mean and standard deviation.

Usage

dgamma2(x, mu = 1, sigma = 1, log = FALSE)

pgamma2(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qgamma2(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rgamma2(n, mu = 1, sigma = 1)

Arguments

x, q

vector of quantiles

mu

mean parameter, must be positive scalar.

sigma

standard deviation parameter, must be positive scalar.

log, log.p

logical; if TRUE, probabilities/ densities \(p\) are returned as \(\log(p)\).

lower.tail

logical; if TRUE, probabilities are \(P[X <= x]\), otherwise, \(P[X > x]\).

p

vector of probabilities

n

number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dgamma2 gives the density, pgamma2 gives the distribution function, qgamma2 gives the quantile function, and rgamma2 generates random deviates.

Details

This implementation allows for automatic differentiation with RTMB.

Examples

x = rgamma2(1)
d = dgamma2(x)
p = pgamma2(x)
q = qgamma2(p)