Monte Carlo version of sdreport

After optimization of an AD model, sdreportMC can be used to calculate samples of confidence intervals of all model parameters and REPORT()ed quantities including non linear functions of random effects and parameters.

Usage

sdreportMC(
  obj,
  what,
  Hessian = NULL,
  CI = FALSE,
  n = 1000,
  probs = c(0.025, 0.975)
)

Arguments

obj

Object returned by MakeADFun() after optimization

what

Vector of strings with names of parameters and REPORT()ed quantities to be reported

Hessian

Optional Hessian matrix. If not provided, it will be computed from the object

CI

Logical. If TRUE, only confidence intervals instead of samples will be returned

n

Number of samples to draw from the multivariate normal distribution of the MLE

probs

Vector of probabilities for the confidence intervals (ignored if no CIs are computed)

Value

Named list corresponding to the elements of what. Each element has the structure of the corresponding quantity with an additional dimension added for the samples. For example, if a quantity is a vector, the list contains a matrix. If a quantity is a matrix, the list contains an array.

Examples

# currently no examples