Compute the stationary distribution of a homogeneous Markov chain

A homogeneous, finite state Markov chain that is irreducible and aperiodic converges to a unique stationary distribution, here called \(\delta\). As it is stationary, this distribution satisfies $$\delta \Gamma = \delta,$$ subject to \(\sum_{j=1}^N \delta_j = 1\), where \(\Gamma\) is the transition probability matrix. This function solves the linear system of equations above.

Usage

stationary(Gamma)

Arguments

Gamma

transition probability matrix of dimension c(N,N)

Value

stationary distribution of the Markov chain with the given transition probability matrix

Examples

Gamma = tpm(c(rep(-2,3), rep(-3,3)))
delta = stationary(Gamma)