Build the prediction design matrix based on new data and model_matrices object created by make_matrices

Build the prediction design matrix based on new data and model_matrices object created by make_matrices

Usage

pred_matrix(model_matrices, newdata)

Arguments

model_matrices

model_matrices object as returned from make_matrices

newdata

data frame containing the variables in the formula and new data for which to evaluate the basis

Value

prediction design matrix for newdata with the same basis as used for model_matrices

Examples

modmat = make_matrices(~ s(x), data.frame(x = 1:10))
Z_predict = pred_matrix(modmat, data.frame(x = 1:10 - 0.5))